feat: Methods to call the breed webservices
A new utils file has been created containing a few functions that will
make it possible to fetch information from the external webservice from
which we'll extract the data to populate the app.

A couple of sample json files have been added to the public folder in
order to make local tests for the time being, since CORS errors are been thrown when trying
the app locally.