style: Base layout for the main component
Although it's still temporary and we have a lot of work yet to do, a
couple of styles has been added to the page by including a small logo at
the top and wrapping the page contents in a flex container.

Standalone components must be created at least for the following items:

- Breed selector
- Selected breed gallery

Also, it will be desirable to include an standalone component for the
header with the logo, as well as a loader that could be displayed
whenever a request to the server is been performed.