refactor: Breed selector logic moved to component
Similarly to what was done in the previous commit, the logic in charge
of handling the breed selector has been rellocated to its component.

The main component keeps on its state a single property with the
selected breed, which is the only data that will be shared between the
components. The file has been greatly simplified, containing nothing
more than the layout of the page.