refactor: Generic HTTP methods moved to a new file
A new file has been created in the utils folder to contain a set of
methods that can easily be reused in different projects.

The method
"fetchJson" matches this description, as it parses a server response
into a Json object and checks whether an HTTP error has ocurred or not. If an HTTP error occurs, an exception will be thrown, which is not
something that happens by default.

This method is now being used in the
"DogAPI" file, which now contains only code specific to the API we're
using for this app.