fix: Update selected breed in the UI
There was an issue with the selected breed, having inconsistant values
in both BreedSelector and App components. The selector was calling the
parent component callback so its state was updated with the selected
breed, but it did not modify its own state. This caused the select box
to always have the placeholder as its selected option, regardless of the
app being aware of the real value.

In order to avoid further inconsistencies, the BreedSelector relies on
the parent to pass it the breed it must highlight by prop, rather than
keeping a duplicate variable on its state.

Redux could help with this, but given the scope of the project, adding
such dependency should not be necessary.